Move resolveMentionSuggestions to createClient
