+++ packages/liveblocks-core/src/types/OptionalPromise.ts
@@ -1,0 +1,1 @@
+export type OptionalPromise<T> = T | Promise<T>;
