+++ packages/liveblocks-react/src/index.ts
@@ -6,11 +6,7 @@ detectDupes(PKG_NAME, PKG_VERSION, PKG_FORMAT);
 
 export { ClientSideSuspense } from "./ClientSideSuspense";
 export { createRoomContext, useRoomContextBundle } from "./room";
-export type {
-  MutationContext,
-  ResolveMentionSuggestionsArgs,
-  ResolveUsersArgs,
-} from "./types";
+export type { MutationContext, ResolveUsersArgs } from "./types";
 
 // Re-exports from @liveblocks/client, for convenience
 export type { Json, JsonObject } from "@liveblocks/client";
